+++ km3pipe/tools.py
@@ -13,6 +13,7 @@ import json
 import os
 import re
 import requests
+import socket
 import subprocess
 import sys
 import warnings
@@ -437,3 +438,14 @@ def is_coherent(seq):
     ```
     """
     return np.array_equal(seq, range(seq[0], int(seq[-1] + 1)))
+
+
+class AnyBar():
+    """A lightweight interface to the AnyBar macOS app."""
+    def __init__(self, port=1738, address='localhost'):
+        self.port = port
+        self.address = address
+        self.sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+
+    def change(self, color):
+        self.sock.sendto(color.encode('utf-8'), (self.address, self.port))